wcftestingtool
It is software made in Microsoft .net WPF, WCF and C#. It takes input as a wcf service base address. As output, it lists all the methods of the service and on selection of any method it asks for the method's parameters (if any). Once we enter the parameters and click the run button, it shows the result of the methods on right panel.